requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
absl-py=0.15.0=pypi_0
affine=2.3.1=pypi_0
aiocontextvars=0.2.2=pypi_0
antlr4-python3-runtime=4.8=pypi_0
anykeystore=0.2=pypi_0
apex=0.1=pypi_0
astor=0.8.1=pypi_0
astunparse=1.6.3=pypi_0
attrs=21.4.0=pypi_0
blas=1.0=mkl
blessings=1.7=pypi_0
bzip2=1.0.8=h7b6447c_0
ca-certificates=2022.9.24=ha878542_0
cached-property=1.5.2=pypi_0
cachetools=4.2.4=pypi_0
certifi=2021.5.30=py36h06a4308_0
chardet=4.0.0=pypi_0
charset-normalizer=2.0.12=pypi_0
click=8.0.4=pypi_0
click-plugins=1.1.1=pypi_0
cligj=0.7.2=pypi_0
colorful=0.5.4=pypi_0
contextvars=2.4=pypi_0
cryptacular=1.6.2=pypi_0
cudatoolkit=11.3.1=h2bc3f7f_2
cycler=0.11.0=pyhd3eb1b0_0
cython=0.29.28=pypi_0
dataclasses=0.8=pyh4f3eec9_6
dbus=1.13.18=hb2f20db_0
decorator=4.4.2=pypi_0
defusedxml=0.7.1=pypi_0
deprecated=1.2.13=pypi_0
docker-pycreds=0.4.0=pypi_0
docopt=0.6.2=pypi_0
emoji=1.7.0=pypi_0
et-xmlfile=1.1.0=pypi_0
expat=2.4.4=h295c915_0
faiss=1.5.3=pypi_0
faiss-cpu=1.7.2=py3.6_h2a577fa_0_cpu
fast-pytorch-kmeans=0.1.6=pypi_0
fastcluster=1.2.6=pypi_0
fasttext=0.9.2=pypi_0
ffmpeg=4.3=hf484d3e_0
filelock=3.4.1=pypi_0
fire=0.4.0=pypi_0
flatbuffers=1.12=pypi_0
fontconfig=2.13.1=h6c09931_0
freetype=2.11.0=h70c0345_0
future=0.18.2=pypi_0
gast=0.2.2=pypi_0
gensim=3.8.3=pypi_0
glib=2.69.1=h4ff587b_1
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
google-auth=1.35.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
google-pasta=0.2.0=pypi_0
gpustat=0.6.0=pypi_0
greenlet=1.1.2=pypi_0
grpcio=1.34.1=pypi_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
h5py=2.10.0=pypi_0
hdbscan=0.8.27=pypi_0
huggingface-hub=0.4.0=pypi_0
hupper=1.10.3=pypi_0
icu=58.2=he6710b0_3
idna=3.3=pypi_0
image-similarity-measures=0.3.5=pypi_0
imagehash=4.2.1=pypi_0
imageio=2.15.0=pypi_0
imbalanced-learn=0.8.1=pypi_0
imblearn=0.0=pypi_0
immutables=0.18=pypi_0
importlib-metadata=4.8.3=pypi_0
importlib-resources=5.4.0=pypi_0
install=1.3.5=pypi_0
intel-openmp=2022.0.1=h06a4308_3633
joblib=1.0.1=pyhd3eb1b0_0
jpeg=9d=h7f8727e_0
jsonlines=3.0.0=pypi_0
keras=2.3.1=pypi_0
keras-applications=1.0.8=pypi_0
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.3.1=py36h2531618_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libblas=3.9.0=1_h6e990d7_netlib
libcblas=3.9.0=3_h893e4fe_netlib
libfaiss=1.7.2=h2bc3f7f_0_cpu
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libgomp=9.3.0=h5101ec6_17
libiconv=1.15=h63c8f33_5
libidn2=2.3.2=h7f8727e_0
liblapack=3.9.0=3_h893e4fe_netlib
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.3.0=hd4cf53a_17
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=h85742a9_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.0.3=h7f8727e_2
libuv=1.40.0=h7b6447c_0
libwebp-base=1.2.2=h7f8727e_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.12=h03d6c58_0
loguru=0.5.1=pypi_0
lz4-c=1.9.3=h295c915_1
markdown=3.3.6=pypi_0
markupsafe=2.0.1=pypi_0
matplotlib=3.3.4=py36h06a4308_0
matplotlib-base=3.3.4=py36h62a2d02_0
mkl=2020.2=256
mkl-service=2.3.0=py36he8ac12f_0
mkl_fft=1.3.0=py36h54f3939_0
mkl_random=1.1.1=py36h0573a6f_0
munch=2.5.0=pypi_0
ncurses=6.3=h7f8727e_2
nettle=3.7.3=hbbd107a_1
networkx=2.5.1=pypi_0
nlg-eval=2.3=pypi_0
nltk=3.6.7=pypi_0
numpy=1.18.5=pypi_0
nvidia-cublas=11.2.1.74=pypi_0
nvidia-cuda-cupti=11.1.69=pypi_0
nvidia-cuda-nvcc=11.1.74=pypi_0
nvidia-cuda-nvrtc=11.1.105=pypi_0
nvidia-cuda-runtime=11.1.74=pypi_0
nvidia-cudnn=8.0.4.30=pypi_0
nvidia-cufft=10.3.0.74=pypi_0
nvidia-curand=10.2.2.74=pypi_0
nvidia-cusolver=11.0.0.74=pypi_0
nvidia-cusparse=11.2.0.275=pypi_0
nvidia-dali-cuda110=0.27.0=pypi_0
nvidia-dali-nvtf-plugin=0.27.0+nv20.11=pypi_0
nvidia-ml-py3=7.352.0=pypi_0
nvidia-nccl=2.8.2=pypi_0
nvidia-pyindex=1.0.9=pypi_0
nvidia-tensorboard=1.15.0+nv20.11=pypi_0
nvidia-tensorflow=1.15.4+nv20.11=pypi_0
nvidia-tensorrt=7.2.1.6=pypi_0
oauthlib=3.2.0=pypi_0
olefile=0.46=py36_0
opencv-python=4.6.0.66=pypi_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h3ad879b_0
openpyxl=3.0.10=pypi_0
openssl=1.1.1q=h7f8727e_0
opt-einsum=3.3.0=pypi_0
packaging=21.3=pypi_0
pandas=1.1.5=py36ha9443f7_0
pastedeploy=2.1.1=pypi_0
pathtools=0.1.2=pypi_0
pbkdf2=1.3=pypi_0
pcre=8.45=h295c915_0
phasepack=1.5=pypi_0
pillow=8.3.1=py36h2c7a002_0
pip=21.2.2=py36h06a4308_0
piqa=1.2.2=pypi_0
plaster=1.0=pypi_0
plaster-pastedeploy=0.7=pypi_0
plotly=5.6.0=pyhd3eb1b0_0
pretrainedmodels=0.7.4=pypi_0
promise=2.3=pypi_0
protobuf=3.19.4=pypi_0
psutil=5.9.0=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pybind11=2.9.2=pypi_0
pyfftw=0.12.0=pypi_0
pynvml=11.4.1=pypi_0
pyparsing=3.0.4=pyhd3eb1b0_0
pyqt=5.9.2=py36h05f1152_2
pyramid=2.0=pypi_0
pyramid-mailer=0.15.1=pypi_0
python=3.6.13=h12debd9_1
python-dateutil=2.8.2=pyhd3eb1b0_0
python3-openid=3.2.0=pypi_0
pytorch-msssim=0.2.1=pypi_0
pytz=2021.3=pyhd3eb1b0_0
pywavelets=1.1.1=pypi_0
pywsd=1.2.4=pypi_0
pyyaml=6.0=pypi_0
qt=5.9.7=h5867ecd_1
rasterio=1.2.10=pypi_0
readline=8.1.2=h7f8727e_1
regex=2022.3.15=pypi_0
repoze-sendmail=4.4.1=pypi_0
requests=2.27.1=pypi_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.8=pypi_0
sacremoses=0.0.49=pypi_0
scikit-image=0.17.2=pypi_0
scikit-learn=0.24.2=py36ha9443f7_0
scipy=1.4.1=pypi_0
seaborn=0.11.2=pypi_0
sentry-sdk=1.9.0=pypi_0
setproctitle=1.2.3=pypi_0
setuptools=58.0.4=py36h06a4308_0
shortuuid=1.0.9=pypi_0
sip=4.19.8=py36hf484d3e_0
six=1.15.0=pypi_0
smart-open=6.0.0=pypi_0
snuggs=1.4.7=pypi_0
some-package=0.1=pypi_0
sqlalchemy=1.4.36=pypi_0
sqlite=3.37.2=hc218d9a_0
stanfordnlp=0.2.0=pypi_0
stanza=1.4.0=pypi_0
tenacity=8.0.1=py36h06a4308_0
tensorboard=1.15.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tensorflow-estimator=1.15.1=pypi_0
termcolor=1.1.0=pypi_0
theano=1.0.5=pypi_0
threadpoolctl=2.2.0=pyh0d69192_0
tifffile=2020.9.3=pypi_0
tk=8.6.11=h1ccaba5_0
tokenizers=0.10.3=pypi_0
tomli=1.2.3=pypi_0
torch=1.10.2+cu113=pypi_0
torchvision=0.11.3+rocm4.2=pypi_0
tornado=6.1=py36h27cfd23_0
tqdm=4.62.3=pyhd3eb1b0_1
transaction=3.0.1=pypi_0
transformers=4.11.0=pypi_0
translationstring=1.4=pypi_0
typing-extensions=3.7.4.3=pypi_0
typing_extensions=3.10.0.2=pyh06a4308_0
urllib3=1.26.9=pypi_0
velruse=1.1.1=pypi_0
venusian=3.0.0=pypi_0
wandb=0.12.21=pypi_0
webencodings=0.5.1=pypi_0
webob=1.8.7=pypi_0
werkzeug=2.0.3=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
wn=0.0.23=pypi_0
wrapt=1.12.1=pypi_0
wtforms=3.0.0=pypi_0
wtforms-recaptcha=0.3.2=pypi_0
xdg=5.1.1=pypi_0
xz=5.2.5=h7b6447c_0
zipp=3.6.0=pypi_0
zlib=1.2.11=h7f8727e_4
zope-deprecation=4.4.0=pypi_0
zope-interface=5.4.0=pypi_0
zope-sqlalchemy=1.6=pypi_0
zstd=1.4.9=haebb681_0
